requirements.txt
streamlit==1.46.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
openai==1.30.1
pandas==2.2.2
pytest==7.4.0
black==23.7.0
flake8==6.0.0
pre-commit==3.3.3
